Organized 'users' functionality vertically.
