+++ opensaas-sh/app_diff/src/client/landing-page/LandingPage.tsx.diff
@@ -20,10 +20,10 @@
 +  footerNavigation,
 +  testimonials,
 +} from './contentSections';
- import DropdownUser from '../components/DropdownUser';
+ import DropdownUser from '../../user/DropdownUser';
+ import { UserMenuItems } from '../../user/UserMenuItems';
 -import { DocsUrl } from '../../common';
 +import { DocsUrl, GithubUrl } from '../../common';
- import { UserMenuItems } from '../components/UserMenuItems';
  import DarkModeSwitcher from '../admin/components/DarkModeSwitcher';
  
  export default function LandingPage() {
