+++ opensaas-sh/app_diff/src/user/operations.ts.diff
@@ -1,6 +1,6 @@
---- template/app/src/server/queries.ts
-+++ opensaas-sh/app/src/server/queries.ts
-@@ -110,6 +110,7 @@
+--- template/app/src/user/operations.ts
++++ opensaas-sh/app/src/user/operations.ts
+@@ -80,6 +80,7 @@
              mode: 'insensitive',
            },
            isAdmin: args.isAdmin,
@@ -8,7 +8,7 @@
          },
          {
            OR: [
-@@ -150,6 +151,7 @@
+@@ -120,6 +121,7 @@
              mode: 'insensitive',
            },
            isAdmin: args.isAdmin,
