+++ template/app/src/client/landing-page/LandingPage.tsx
@@ -8,9 +8,9 @@ import { BiLogIn } from 'react-icons/bi';
 import logo from '../static/logo.png';
 import openSaasBanner from '../static/open-saas-banner.png';
 import { features, navigation, faqs, footerNavigation, testimonials } from './contentSections';
-import DropdownUser from '../components/DropdownUser';
+import DropdownUser from '../../user/DropdownUser';
+import { UserMenuItems } from '../../user/UserMenuItems';
 import { DocsUrl } from '../../common';
-import { UserMenuItems } from '../components/UserMenuItems';
 import DarkModeSwitcher from '../admin/components/DarkModeSwitcher';
 
 export default function LandingPage() {
