+++ template/app/src/server/queries.ts
@@ -3,10 +3,8 @@ import { HttpError } from 'wasp/server';
 import {
   type GetGptResponses,
   type GetDailyStats,
-  type GetPaginatedUsers,
   type GetAllTasksByUser,
 } from 'wasp/server/operations';
-import { type SubscriptionStatus } from '../payment/plans';
 
 type DailyStatsWithSources = DailyStats & {
   sources: PageViewSource[];
@@ -70,108 +68,4 @@ export const getDailyStats: GetDailyStats<void, DailyStatsValues> = async (_args
   });
 
   return { dailyStats, weeklyStats };
-};
-
-type GetPaginatedUsersInput = {
-  skip: number;
-  cursor?: number | undefined;
-  emailContains?: string;
-  isAdmin?: boolean;
-  subscriptionStatus?: SubscriptionStatus[];
-};
-type GetPaginatedUsersOutput = {
-  users: Pick<
-    User,
-    'id' | 'email' | 'username' | 'lastActiveTimestamp' | 'subscriptionStatus' | 'stripeId'
-  >[];
-  totalPages: number;
-};
-
-export const getPaginatedUsers: GetPaginatedUsers<GetPaginatedUsersInput, GetPaginatedUsersOutput> = async (
-  args,
-  context
-) => {
-  if (!context.user?.isAdmin) {
-    throw new HttpError(401);
-  }
-
-  const allSubscriptionStatusOptions = args.subscriptionStatus as Array<string | null> | undefined;
-  const hasNotSubscribed = allSubscriptionStatusOptions?.find((status) => status === null) 
-  let subscriptionStatusStrings = allSubscriptionStatusOptions?.filter((status) => status !== null) as string[] | undefined
-
-  const queryResults = await context.entities.User.findMany({
-    skip: args.skip,
-    take: 10,
-    where: {
-      AND: [
-        {
-          email: {
-            contains: args.emailContains || undefined,
-            mode: 'insensitive',
-          },
-          isAdmin: args.isAdmin,
-        },
-        {
-          OR: [
-            {
-              subscriptionStatus: {
-                in: subscriptionStatusStrings,
-              },
-            },
-            {
-              subscriptionStatus: {
-                equals: hasNotSubscribed,
-              },
-            },
-          ],
-        },
-      ],
-    },
-    select: {
-      id: true,
-      email: true,
-      username: true,
-      isAdmin: true,
-      lastActiveTimestamp: true,
-      subscriptionStatus: true,
-      stripeId: true,
-    },
-    orderBy: {
-      id: 'desc',
-    },
-  });
-
-  const totalUserCount = await context.entities.User.count({
-    where: {
-      AND: [
-        {
-          email: {
-            contains: args.emailContains || undefined,
-            mode: 'insensitive',
-          },
-          isAdmin: args.isAdmin,
-        },
-        {
-          OR: [
-            {
-              subscriptionStatus: {
-                in: subscriptionStatusStrings,
-              },
-            },
-            {
-              subscriptionStatus: {
-                equals: hasNotSubscribed,
-              },
-            },
-          ],
-        },
-      ],
-    },
-  });
-  const totalPages = Math.ceil(totalUserCount / 10);
-
-  return {
-    users: queryResults,
-    totalPages,
-  };
 };
